feat: PostHeader.tsx
