+++ app/_components/PostViewer.tsx
@@ -1,30 +1,12 @@
 import PostContent from './PostContent';
+import PostHeader from './PostHeader';
 
-export interface Post {
-  title: string;
-  author: string;
-  tags: string[];
-  content: string;
-}
+import { Post } from '../../lib/PostManager';
 
 export default function PostViewer({ post }: { post: Post }) {
   return (
     <div className="post-wrapper">
-      <div className="post-title">{post.title}</div>
-      <div className="post-info">
-        <div className="post-tags">
-          {post.tags
-            ? post.tags.map((tag) => (
-                <div className="post-tag" key={tag}>
-                  {tag}
-                </div>
-              ))
-            : null}
-        </div>
-      </div>
-      <div className="post-author">
-        ✍️ written by <strong>{post.author}</strong>
-      </div>
+      <PostHeader data={post.frontMatter} />
       <PostContent content={post.content} />
     </div>
   );
